:recycle: separate popup from service
